+++ spec/spec_helper.rb
@@ -3,7 +3,10 @@ $LOAD_PATH.unshift(File.join(File.dirname(__FILE__), '..', 'lib'))
 require 'rubygems'
 require 'rspec'
 require 'graphlient'
-require 'byebug' if RUBY_ENGINE != 'jruby'
+begin
+  require 'byebug' if RUBY_ENGINE != 'jruby'
+rescue LoadError # byebug not available on all platforms (e.g. Windows x64-mingw-ucrt)
+end
 require 'rack/test'
 require 'webmock/rspec'
 require 'vcr'